Add the ability to skip CRC and Now Loading to webapp
This allows the user to skip CRC checking and the Now Loading screen when using the web application. Adds data to cabinet config. This is how I got the software booting on Chihiro hardware. I suspect an issue with the CRC checking function, and I will be diving more into this in the future. Quick and simple fix. Expect bugs.